update: code compress for iosscroll
