+++ frontend/kezuler-fe/src/components/my-page/MyPageEdit.tsx
@@ -17,6 +17,7 @@ import { usePatchUser } from 'src/hooks/usePatchUser';
 import { alertAction } from 'src/reducers/alert';
 import { AppDispatch } from 'src/store';
 import getCurrentUserInfo from 'src/utils/getCurrentUserInfo';
+import { focusDisable, focusEnable } from 'src/utils/iosScrollDisable';
 
 import BottomButton from '../common/BottomButton';
 
@@ -58,31 +59,12 @@ function MyPageEdit({ goToMain }: Props) {
   const onFocus = () => setFocused(true);
   const onBlur = () => setFocused(false);
 
-  function preventIOSScroll() {
-    const focusedInput = document.activeElement as
-      | HTMLInputElement
-      | HTMLTextAreaElement;
-    focusedInput?.blur();
-  }
-
-  function disable() {
-    document
-      .querySelector('.App')
-      ?.addEventListener('touchmove', preventIOSScroll);
-  }
-
-  function enable() {
-    document
-      .querySelector('.App')
-      ?.removeEventListener('touchmove', preventIOSScroll);
-  }
-
   useEffect(() => {
     if (isMobile && isIOS) {
       if (focused) {
-        disable();
+        focusDisable();
       } else {
-        enable();
+        focusEnable();
       }
     }
   }, [focused]);
